requirements.txt
alembic==1.12.0
blinker==1.6.2
click==8.3.1
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.3.3
Flask-Admin==1.6.1
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.1.1
greenlet==2.0.2
gunicorn==21.2.0
idna==3.6
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
packaging==24.0
pandas
psycopg2==2.9.7
python-dotenv==1.0.0
pytz==2023.3.post1
SQLAlchemy==2.0.20
typing_extensions==4.13.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.16.2
